split up sequences from graph
