+++ scripts/diff2vec/run_diff2vec.py
@@ -1,17 +1,10 @@
-import pandas as pd
-from typing import Any, List, Tuple
-
-from src.diff2vec.graph import UndirectedGraph
+from typing import Any, List
+from src.utils.utils import from_pickle
 from src.diff2vec.diff2vec import Diff2Vec
 
 
 def main(args: Any):
-    print('Loading data from CSV')
-    data: pd.DataFrame = pd.read_csv(args.data_csv)
-    print('Building graph')
-    graph: UndirectedGraph = build_graph(data)
-    print(f'Made graph with {len(graph)} nodes')
-
+    sequences: List[List[int]] = from_pickle(args.sequences_file)
     model: Diff2Vec = Diff2Vec(
         dimensions = args.dim,
         window_size = args.window,
@@ -22,28 +15,13 @@ def main(args: Any):
         seed = args.seed,
         cache_dir = args.cache_dir,
     )
-    model.fit(graph)
-
-
-def build_graph(data: pd.DataFrame) -> UndirectedGraph:
-    node_a: List[int] = data.from_address.to_numpy().tolist()
-    node_b: List[int] = data.to_address.to_numpy().tolist()
-    edge_ab: List[Tuple[int, int]] = list(zip(node_a, node_b))
-
-    graph: UndirectedGraph = UndirectedGraph()
-    graph.add_nodes_from(node_a)
-    graph.add_nodes_from(node_b)
-    graph.add_edges_from(edge_ab)
-
-    return graph
-
+    model.fit(sequences)
 
 
 if __name__ == "__main__":
     from argparse import ArgumentParser
     parser: ArgumentParser = ArgumentParser()
-    parser.add_argument('data_csv', type=str, help='path to save data')
-    parser.add_argument('cache_dir', type=str, help='path to cache')
+    parser.add_argument('sequences_file', type=str, help='path to cached sequences')
     parser.add_argument('--epochs', type=int, default=10, help='epochs (default: 10)')
     parser.add_argument('--workers', type=int, default=4, help='workers (default: 4)')
     parser.add_argument('--lr', type=float, default=0.05, help='learning rate (default: 0.05)')
